persistence/requirements.txt
# CWB Hub Persistence System Requirements

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Authentication
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# FastAPI integration
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Validation
pydantic>=2.5.0
email-validator>=2.1.0

# Utilities
python-dotenv>=1.0.0
redis>=5.0.0
celery>=5.3.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-postgresql>=5.0.0
factory-boy>=3.3.0

# Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0